Initial conversion to Rollup bundling with CommonJS, ESM/MJS, and AMD format output
