+++ LICENSE.txt
@@ -1,6 +1,7 @@
 The MIT License (MIT)
 
 Copyright (c) 2015 Matt Asher (http://Mattasher.com)
+Copyright (c) 2018 Shane Holloway (http://ShaneHolloway.com) for Rollup-based packaging
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
