+++ package.json
@@ -2,8 +2,13 @@
   "name": "probability-distributions",
   "version": "0.9.1",
   "description": "Generate random variables from a variety of probability distributions. Includes tools to shuffle an array or sample from it.",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "module": "dist/probability-distributions.esm.js",
+  "browser": "dist/probability-distributions.amd.js",
   "scripts": {
+    "build": "rollup --config",
+    "watch": "npm -s run build -- --watch",
+    "pretest": "npm -s run build",
     "test": "mocha"
   },
   "keywords": [
@@ -31,15 +36,13 @@
   ],
   "author": "Matt Asher",
   "license": "MIT",
-  "dependencies": {
-    "crypto": "0.0.3"
-  },
   "repository": {
     "type": "git",
     "url": "https://github.com/Mattasher/probability-distributions"
   },
   "devDependencies": {
     "chai": "^3.3.0",
-    "mocha": "^2.3.3"
+    "mocha": "^2.3.3",
+    "rollup": "^0.55.3"
   }
 }
